Hi,this is the pr for [issue](https://github.com/hbollon/go-edlib/issues/16).
I run the benchmark on OSADamerauLevenshtein.
Here is the result.
| name | before | after | alloc_before | alloc_after | 
| :-----:| :----: | :----: | :----: | :----: |
| sameLengthStringInput | 3389 ns/op |2416 ns/op |23 allocs/op |4 allocs/op |
| pneumonoultramicroscopi | 15462 ns/op|  10253 ns/op |51 allocs/op|  8 allocs/op |
|こにんちこにんちこに | 4027 ns/op |  3342 ns/op |18 allocs/op | 4 allocs/op|
| I_love_horror_movi |  4282 ns/op |  3073 ns/op | 22 allocs/op |  4 allocs/op |